ai_server/requirements.txt
flask==2.3.3
pymysql==1.1.0
python-dotenv==1.0.0
google-generativeai>=0.7.0
flask-cors==4.0.0